chore: rename cdn server package to rpcserver
