+++ cdnsystem/rpcserver/rpcserver.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package server
+package rpcserver
 
 import (
 	"context"
