+++ cdnsystem/rpcserver/rpcserver_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package server
+package rpcserver
 
 import (
 	"context"
